% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zen_collection_list.R
\name{zen_collections}
\alias{zen_collections}
\title{List Zenodo products}
\usage{
zen_collections(access_token = getOption("zenodo_token"))
}
\arguments{
\item{acces_token}{Your Zenodo access token}
}
\description{
List all depositions for the currently authenticated user.
}

