% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zen_collection_create.R
\name{zen_create}
\alias{zen_create}
\title{Create a collection on Zenodo}
\usage{
zen_create(collection_name = NULL, access_token = getOption("zenodo_token"))
}
\arguments{
\item{collection_name}{Name of a new collection to be created on Zenodo}

\item{access_token}{Your Zenodo access token}
}
\description{
Create a collection on Zenodo
}

