% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zenodo_dir.R
\name{zen_dir}
\alias{zen_dir}
\title{List Zenodo products}
\usage{
zen_dir(access_token = getOption("zenodo_token"))
}
\arguments{
\item{acces_token}{Your Zenodo access token}
}
\description{
List Zenodo products
}

