% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{zen_download}
\alias{zen_download}
\title{Function to download files from a Zenodo deposit.}
\usage{
zen_download(id, dir_disk, access_token = Sys.getenv("ZENODO_TOKEN"))
}
\arguments{
\item{id}{An id of a Zenodo deposit.}

\item{dir_disk}{Directory on disk where the Zenodo files should be put.}

\item{access_token}{An individual access_token.}
}
\value{
An API response as a list.
}
\description{
Use this function to download files from a Zenodo deposit.
}
\examples{
\dontrun{
# Function needs a personal API key.
id = 5947275
dir_disk = '~/Desktop/look'
zen_download(id, dir_disk)
}

}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
