% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zen_fetch.R
\name{zen_fetch}
\alias{zen_fetch}
\title{http GET or POST request to Zenodo server using appropriate authentication}
\usage{
zen_fetch(path, body = NULL, parse.json = TRUE, include_headers = FALSE,
  zen_auth_params = list(), parse_params = list(), ...)
}
\arguments{
\item{path}{The path on the zenodo server relative to the zenodo root}

\item{body}{The (optional) body of the post request, usually in the form of a
named list. See the \code{\link[httr]{POST}} documentation for full
details.}

\item{parse.json}{Whether or not to parse a JSON response to an R object
(default \code{TRUE})}

\item{include_headers}{Whether to include basic headers from the http request
as attributes on the parsed JSON object (default \code{FALSE}).}

\item{zen_auth_params}{Named list of authentication parameters passed to
zen_auth to override default authetication process.}

\item{parse_params}{Named list of parameters passed to
\code{httr::\link[httr]{content}} function and then most likely on to
\code{jsonlite::\link[jsonlite]{fromJSON}}.}

\item{...}{Additional arguments passed to the \code{httr::GET} or
\code{httr::POST} function}
}
\value{
When \code{parse.json=FALSE} an object of class \code{response}
  otherwise the raw R object generated by calling \code{content(response,
  as='parsed')} on the body of the response.
}
\description{
\code{zen_fetch} carries out a GET operation when
  \code{body=NULL}, POST otherwise. The http status code of the response will
  be checked - if invalid an error will be thrown.
}
\details{
Authentication is presently handled by the \code{\link{zen_pat}}
  function.
}
\examples{
\dontrun{
# Get list of depositions for current user
zen_fetch('api/deposit/depositions')

# Get jsonlite to do some extra work on the result
# see ?fromJSON for details
zz=zen_fetch('api/deposit/depositions', parse_params = list(simplifyVector=TRUE))

## You can also PUT
# Here we take a sample JSON request  from the Zenodo API reference
# https://zenodo.org/dev#restapi-res-dep
json_request=paste0('{"metadata": {"title": "My first upload", "upload_type": "poster",',
  ' "description": "This is my first upload", "creators": [{"name": "Doe, John",',
  ' "affiliation": "Zenodo"}]}}')
# see what corresponding R list would look like
dput(fromJSON(d), control='useSource' )

# now use that list as the body for a PUT
body=list(metadata = list(title = "My first upload", upload_type = "poster",
  description = "This is my first upload", creators = list(name = "Doe, John",
    affiliation = "Zenodo")))
zen_fetch('api/deposit/depositions', body=body)
}
}
\seealso{
\code{\link{zen_pat}}, \code{\link[httr]{GET}},
  \code{\link[httr]{POST}}, \code{\link[jsonlite]{fromJSON}}
}

