% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zen_auth.R
\name{zen_pat}
\alias{zen_pat}
\title{Get Zenodo Personal Access Token (PAT)}
\usage{
zen_pat()
}
\description{
This will first check for R option \code{zenodo_token}. If that
  does not exist then the environment variable ZENODO_PAT will be checked.
}
\details{
You can make a new Zenodo Personal Access Token at
  \url{https://zenodo.org/account/settings/applications/}.
}

