% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{zen_put}
\alias{zen_put}
\title{Function to put a file into a Zenodo Upload.}
\usage{
zen_put(id, file_name, file_disk, access_token = Sys.getenv("ZENODO_TOKEN"))
}
\arguments{
\item{id}{An id of a Zenodo Upload.}

\item{file_name}{Name of the file which it should have on Zenodo.}

\item{file_disk}{A file on disk tp upload.}

\item{access_token}{An individual access_token.}
}
\value{
An API response as a list.
}
\description{
Use this function to put a file into a Zenodo upload.
}
