% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{zen_upload}
\alias{zen_upload}
\title{Create a Zenodo Draft deposit.}
\usage{
zen_upload(
  title = "My Title",
  upload_type = "dataset",
  publication_type = NULL,
  publication_date = NULL,
  author = NULL,
  description = NULL,
  version = NULL,
  language = NULL,
  keywords = NULL,
  notes = NULL,
  access_right = "open",
  access_right_addition = NULL,
  access_token = Sys.getenv("ZENODO_TOKEN")
)
}
\arguments{
\item{title}{Title of deposit.}

\item{upload_type}{Type of deposit. Can be one of: 'publication', 'poster',
'presentation', 'dataset', 'image', 'video', 'software', 'lesson', 'physicalobject',
'other'.}

\item{publication_type}{Not yet documented.}

\item{publication_date}{Date of publication. Submit a object of class Date, IDate
or a date string fromated as 'YYYY-MM-DD'.}

\item{author}{A list of lists containing name, affiliation and ORCID of the author.
Multiple entries are possible.}

\item{description}{Description of the data as a string.}

\item{version}{A version string. Semantically-versioned tag is recommended:
\url{https://semver.org/}.}

\item{language}{A lower-cased 3-letter ISO 639-3 string (e.g. 'ger').}

\item{keywords}{Keywords as a vector or string.}

\item{notes}{Additional notes as a string.}

\item{access_right}{Has to be one of 'open', 'embargoed', 'restricted', 'closed'.}

\item{access_right_addition}{Only needs to be specified if 'embargoed' or
'restricted' is selected in access_right. For 'embargoed' provide a date string
('YYYY-MM-DD'). For 'restricted' provide a text string explaining the conditions.}

\item{access_token}{An individual access_token.}
}
\value{
An API response as a list.
}
\description{
Creates a new draft deposit to Zenodo. Files can be added with
the help of the zen_bucket() function. If you want to publish the
deposit, please do this manually on Zenodo.
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
