% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zenodo-package.R
\docType{package}
\name{zenodo-package}
\alias{zenodo}
\alias{zenodo-package}
\title{Programmatic access to the Zenodo API}
\description{
See \url{https://zenodo.org/dev} for details
}
\section{Authentication}{
 At present the only functional authentication scheme
  is the use of a Zenodo Personal Access Token. See \code{\link{zen_pat}} for
  details.
}
\keyword{package}

