% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zen_collection_list.R
\name{zen_collections}
\alias{zen_collections}
\title{List Zenodo collections.}
\usage{
zen_collections(access_token = zen_pat())
}
\arguments{
\item{access_token}{Your Zenodo access token}
}
\description{
List all depositions for the currently authenticated user. Collections
typically archive all outputs of a research project.
}
\examples{
\dontrun{
my_collections <- zen_collections()
}
}

